split validation_test.go
